pyproject.toml
[project]
name = "slack-summary-bot"
version = "0.1.0"
description = "Slack 대화 요약 봇"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "slack-sdk",
    "python-dotenv",
    "openai",
    "schedule"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
